feat(rpc): service-layer overhaul with TLS and API-key auth
- new rpc/service layer: typed results, GetValidator/GetStateRoot/
  GetPendingTxs handlers, ValidatorResult with votingPower and
  uint16 commission
- rpc/middleware: API-key auth (constant-time compare) and security
  headers (nosniff, HSTS when TLS, frame/XSS protection)
- rpc/router + server: RouterConfig (TLS/API key), SecureHeaders,
  backward-compatible constructors, ServeTLS, new dsn_getValidator
  method; legacy placeholders (getStateRoot) now hit the service
- node/config.go: TLSCertFile/TLSKeyFile/RPCApiKey/RPCApiKeyHeader
  fields + env vars (DSN_TLS_*, DSN_RPC_API_KEY*); BlockTimeSec
  rides along (used by the economics commit)
- wire-ups: cmd/dsn/main.go (security-aware RPC startup),
  validator_cmd.go (new dsn_getValidator + response shape),
  explorer (VotingPower/uint16 commission, new RPC method names),
  JS SDK (dsn_getBlock/dsn_getTransaction/dsn_sendTransaction
  object params)

+++ explorer/types.go
@@ -53,7 +53,7 @@ type EventItem struct {
 type ValidatorInfo struct {
 	Address    string `json:"address"`
 	Power      uint64 `json:"power"`
-	Commission uint64 `json:"commission"`
+	Commission uint16 `json:"commission"`
 }
 
 // SupplyInfo represents token supply metrics.
@@ -67,7 +67,7 @@ type SupplyInfo struct {
 type ContractInfo struct {
 	Address      string        `json:"address"`
 	CodeHash     string        `json:"codeHash"`
-	Metadata    *ContractMeta `json:"metadata,omitempty"`
+	Metadata     *ContractMeta `json:"metadata,omitempty"`
 	RecentEvents []EventItem   `json:"recentEvents,omitempty"`
 }
